Convert commands to use destructuring syntax
Over time, there will be a large number of arguments to the run function
that inject various dependencies. Not all commands will use all the
dependencies, so using the destructuring syntax gives an easy way for
command authors to specify only what they need. This is a one time
breaking change that is fixed by this commit.